Fixing foreseen heroku issue
If you deploy this script to heroku and run it continuously, the dyno hours will be exhausted, which is not the problem as the only times we need it to work are 1st and 15th of a month.

But the problem existed in another area, that is the dyno hours resetting time sometimes comes late, not always 00.00 AM GMT. As the script works when midnight of 1st, the dyno hours may come late and it skips the process. That's why I gave it a compensation until 06.00 AM GMT to get the dyno hours properly.